requirements.txt
# Core dependencies for MLOps platform

# ============================================================
# Python version requirement
# ============================================================
# Requires Python 3.10 or higher

# ============================================================
# Data manipulation and scientific computing
# ============================================================
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# ============================================================
# Machine Learning frameworks
# ============================================================
xgboost>=2.0.0,<3.0.0
lightgbm>=4.0.0,<5.0.0
imbalanced-learn>=0.11.0  # For handling class imbalance

# ============================================================
# MLOps tools
# ============================================================
mlflow>=2.8.0,<3.0.0
dvc>=3.0.0,<4.0.0
dvc-s3>=3.0.0  # S3 remote storage support
great-expectations>=0.18.0,<1.0.0
evidently>=0.4.0,<1.0.0

# ============================================================
# API and web frameworks
# ============================================================
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
streamlit>=1.28.0,<2.0.0
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.6  # For file uploads in FastAPI

# ============================================================
# Monitoring and metrics
# ============================================================
prometheus-client>=0.18.0,<1.0.0
prometheus-fastapi-instrumentator>=6.1.0,<7.0.0

# ============================================================
# Database
# ============================================================
sqlalchemy>=2.0.0,<3.0.0
alembic>=1.12.0,<2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
aiosqlite>=0.19.0  # Async SQLite support

# ============================================================
# Reporting and notifications
# ============================================================
jinja2>=3.1.2,<4.0.0
weasyprint>=60.0,<61.0
matplotlib>=3.8.0,<4.0.0
seaborn>=0.13.0,<1.0.0
plotly>=5.17.0,<6.0.0
kaleido>=0.2.1  # For static image export from plotly

# ============================================================
# Utilities
# ============================================================
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.1,<7.0.0
click>=8.1.0,<9.0.0  # CLI tool
rich>=13.6.0,<14.0.0  # Beautiful terminal output
loguru>=0.7.0,<1.0.0  # Better logging
requests>=2.31.0,<3.0.0  # HTTP library
httpx>=0.25.0,<1.0.0  # Async HTTP client
python-dateutil>=2.8.2,<3.0.0
tenacity>=8.2.0  # Retry logic

# ============================================================
# Data validation and quality
# ============================================================
jsonschema>=4.19.0,<5.0.0
cerberus>=1.3.5,<2.0.0

# ============================================================
# File formats
# ============================================================
openpyxl>=3.1.0  # Excel support
xlrd>=2.0.1  # Old Excel format

# ============================================================
# Scheduling
# ============================================================
apscheduler>=3.10.0,<4.0.0  # Job scheduling

# ============================================================
# Security
# ============================================================
cryptography>=41.0.0,<42.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# ============================================================
# Progress bars and UI
# ============================================================
tqdm>=4.66.0,<5.0.0

# ============================================================
# Type checking (optional runtime)
# ============================================================
typeguard>=4.1.0,<5.0.0
